@@ -1,44 +1,44 @@
 {
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "npm run-script build; jest --runInBand",
-    "build": "tsc --build -v",
-    "dev": "nodemon src/app.ts",
-    "test-linux": "rm -r outdir; npm run-script test;"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git@gitlab.stud.idi.ntnu.no:bachelor-g-131/backend.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@types/cors": "^2.8.10",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "mysql2": "^2.2.5",
-    "sequelize": "^6.6.2",
-    "formidable": "^1.2.2",
-    "@types/formidable": "^1.2.1"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.7",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.1.5",
-    "swagger-jsdoc": "6.0.6",
-    "swagger-ui-express": "^4.1.2",
-    "@types/express": "^4.17.11",
-    "@types/node": "^14.14.28",
-    "@types/swagger-jsdoc": "^6.0.0",
-    "@types/swagger-ui-express": "^4.1.2",
-    "@types/validator": "^13.1.3",
-    "@types/bluebird": "^3.5.33",
-    "jest": "^26.6.3",
-    "@types/jest": "26.0.22"
-  }
+    "name": "backend",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "test": "npm run-script build; jest --runInBand",
+        "build": "tsc --build -v",
+        "dev": "nodemon src/app.ts",
+        "test-linux": "rm -r outdir; npm run-script test;"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git@gitlab.stud.idi.ntnu.no:bachelor-g-131/backend.git"
+    },
+    "keywords": [],
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "@types/cors": "^2.8.10",
+        "@types/formidable": "^1.2.1",
+        "cors": "^2.8.5",
+        "dotenv": "^8.2.0",
+        "express": "^4.17.1",
+        "formidable": "^1.2.2",
+        "mysql2": "^2.2.5",
+        "sequelize": "^6.6.2"
+    },
+    "devDependencies": {
+        "@types/bluebird": "^3.5.33",
+        "@types/express": "^4.17.11",
+        "@types/jest": "26.0.22",
+        "@types/node": "^14.14.28",
+        "@types/swagger-jsdoc": "^6.0.0",
+        "@types/swagger-ui-express": "^4.1.2",
+        "@types/validator": "^13.1.3",
+        "jest": "^26.6.3",
+        "nodemon": "^2.0.7",
+        "swagger-jsdoc": "6.0.6",
+        "swagger-ui-express": "^4.1.2",
+        "ts-node": "^9.1.1",
+        "typescript": "^4.5.5"
+    }
 }
