@@ -70,10 +70,7 @@ export default {
     // ],
 
     // An array of file extensions your modules use
-    moduleFileExtensions: [
-        "js",
-        "json"
-    ],
+    moduleFileExtensions: ["js", "json"],
 
     // A map from regular expressions to module names or to arrays of module names that allow to stub out resources with a single module
     // moduleNameMapper: {},
@@ -112,9 +109,7 @@ export default {
     // rootDir: undefined,
 
     // A list of paths to directories that Jest should use to search for files in
-    roots: [
-        "outdir"
-    ],
+    roots: ["outdir"],
 
     // Allows you to use a custom runner instead of Jest's default test runner
     // runner: "jest-runner",
@@ -150,13 +145,11 @@ export default {
     testMatch: [
         "*test*.js",
         "**/__tests__/**/*.[jt]s?(x)",
-        "**/?(*.)+(spec|test).[tj]s?(x)"
+        "**/?(*.)+(spec|test).[tj]s?(x)",
     ],
 
     // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
-    testPathIgnorePatterns: [
-        "/node_modules/"
-    ],
+    testPathIgnorePatterns: ["/node_modules/"],
 
     // The regexp pattern or array of patterns that Jest uses to detect test files
     // testRegex: [],
