@@ -1,13 +1,13 @@
-import {DataTypes, Model, Sequelize} from "sequelize";
+import { DataTypes, Model, Sequelize } from "sequelize";
 
-module.exports=(sequelize:Sequelize)=>{
-    return sequelize.define<Model<any,any>, unknown>("subject_in_article",
+module.exports = (sequelize: Sequelize) => {
+    return sequelize.define<Model<any, any>, unknown>(
+        "subject_in_article",
+        {},
         {
-
-        },{
-            createdAt:false,
-            updatedAt:false,
-            timestamps:false
+            createdAt: false,
+            updatedAt: false,
+            timestamps: false,
         }
     );
-}
+};
