@@ -1,18 +1,21 @@
-import {DataTypes, Model, Sequelize} from "sequelize";
+import { DataTypes, Model, Sequelize } from "sequelize";
 
-module.exports=(sequelize:Sequelize)=>{
-    return sequelize.define<Model<any,any>, unknown>("subject", {
+module.exports = (sequelize: Sequelize) => {
+    return sequelize.define<Model<any, any>, unknown>(
+        "subject",
+        {
             subject_id: {
                 type: DataTypes.STRING(10),
                 primaryKey: true,
-                unique: true
+                unique: true,
             },
             subject_name: {
-                type: DataTypes.STRING
-            }
-        },{
-            createdAt:false,
-            updatedAt:false
+                type: DataTypes.STRING,
+            },
+        },
+        {
+            createdAt: false,
+            updatedAt: false,
         }
     );
-}
+};
